+++ package.json
@@ -32,6 +32,7 @@
     "react-icons": "^5.2.1",
     "react-intersection-observer": "^9.10.2",
     "react-leaflet": "^4.2.1",
+    "react-leaflet-custom-control": "^1.4.0",
     "react-star-ratings": "^2.3.0",
     "react-syntax-highlighter": "^15.5.0",
     "socket.io-client": "^4.7.5",
